Refactor CLI helpers into SRP modules
